wpt_babel/content-security-policy/base-uri/base-uri-allow.sub.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="base-uri {{location[scheme]}}://{{domains[www1]}}:{{ports[http][0]}}/" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
  <script>
   "use strict";

var t = async_test("Check that base URIs can be set if they do not violate the page's policy.");
window.addEventListener('securitypolicyviolation', t.step_func(function (t) {
  assert_unreached('No CSP violation report should have been fired.');
}));
  </script>
  <base href="{{location[scheme]}}://{{domains[www1]}}:{{ports[http][0]}}/"/>
  <script>
   "use strict";

t.step(function () {
  assert_equals(document.baseURI, "{{location[scheme]}}://{{domains[www1]}}:{{ports[http][0]}}/");
  t.done();
});
  </script>
 </head>
 <body>
 </body>
</html>
